requirements.txt
fastembed 
llama-index-vector-stores-chroma
llama-index-llms-mistralai
llama-index-embeddings-fastembed
llama-index-readers-file